[Refactor] Refactor code layout
